move installDependencies and getWelcomeMessage to PackageManagerController
