+++ packages/create-amplify/src/package-manager-controller/npm_package_manager_controller.ts
@@ -1,6 +1,4 @@
 import { execa as _execa } from 'execa';
-import { executeWithDebugLogger } from '../execute_with_logger.js';
-import { DependencyType } from './package_manager_controller_factory.js';
 import { PackageManagerController } from './package_manager_controller.js';
 
 /**
@@ -17,35 +15,4 @@ export class NpmPackageManagerController extends PackageManagerController {
     this.installCommand = 'install';
     this.initDefault = ['init', '--yes'];
   }
-
-  /**
-   * Installs the given package names as devDependencies
-   */
-  installDependencies = async (
-    packageNames: string[],
-    type: DependencyType
-  ): Promise<void> => {
-    const args = [`${this.installCommand}`].concat(...packageNames);
-    if (type === 'dev') {
-      args.push('-D');
-    }
-
-    await executeWithDebugLogger(
-      this.projectRoot,
-      this.executable,
-      args,
-      this.execa
-    );
-  };
-
-  getWelcomeMessage = () => {
-    const cdCommand =
-      process.cwd() === this.projectRoot
-        ? ''
-        : `cd .${this.projectRoot.replace(process.cwd(), '')}; `;
-
-    return `Welcome to AWS Amplify! 
-Run \`${this.binaryRunner} amplify help\` for a list of available commands. 
-Get started by running \`${cdCommand}${this.binaryRunner} amplify sandbox\`.`;
-  };
 }
