+++ packages/create-amplify/src/package-manager-controller/yarn_modern_package_manager_controller.ts
@@ -17,35 +17,4 @@ export class YarnModernPackageManagerController extends PackageManagerController
     this.installCommand = 'add';
     this.initDefault = ['init', '--yes'];
   }
-
-  /**
-   * Installs the given package names as devDependencies
-   */
-  installDependencies = async (
-    packageNames: string[],
-    type: DependencyType
-  ): Promise<void> => {
-    const args = [`${this.installCommand}`].concat(...packageNames);
-    if (type === 'dev') {
-      args.push('-D');
-    }
-
-    await executeWithDebugLogger(
-      this.projectRoot,
-      this.executable,
-      args,
-      this.execa
-    );
-  };
-
-  getWelcomeMessage = () => {
-    const cdCommand =
-      process.cwd() === this.projectRoot
-        ? ''
-        : `cd .${this.projectRoot.replace(process.cwd(), '')}; `;
-
-    return `Welcome to AWS Amplify! 
-Run \`${this.binaryRunner} amplify help\` for a list of available commands. 
-Get started by running \`${cdCommand}${this.binaryRunner} amplify sandbox\`.`;
-  };
 }
